feat: move cache skeleton to own module
