+++ client_new/Cargo.toml
@@ -19,6 +19,7 @@ std         = []
 thiserror   = "1.0.30" 
 log         = "0.4.14"
 zeroize     = { version = "1.5.0", features = ["zeroize_derive"] }
+serde       = { version = "1.0",   features = [ "derive" ] }
 rlu         = { package = "stronghold-rlu", path = "../rlu/"}
 
 [dev-dependencies]
