+++ client_new/src/types.rs
@@ -1,0 +1,8 @@
+// Copyright 2020-2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+// modules
+mod cache;
+
+// re-export imports
+pub use cache::*;
